+++ packages/amplify-cli-npm/run.ts
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+
 import { run } from '.';
 
 run();