+++ packages/amplify-cli/src/__tests__/context-manager.test.ts
@@ -6,25 +6,25 @@ import { constructContext, attachUsageData } from '../context-manager';
 import { Context } from '../domain/context';
 import { PluginInfo } from '../domain/plugin-info';
 import { PluginManifest } from '../domain/plugin-manifest';
-import { UsageData, NoUsageData} from '../domain/amplify-usageData';
+import { UsageData, NoUsageData } from '../domain/amplify-usageData';
 
 jest.mock('../domain/amplify-usageData/', () => ({
   UsageData: {
     Instance: {
       setIsHeadless: jest.fn(),
-      init : jest.fn(),
+      init: jest.fn(),
     },
   },
   NoUsageData: {
-    Instance:  {
-        setIsHeadless: jest.fn(),
-        init : jest.fn(),
+    Instance: {
+      setIsHeadless: jest.fn(),
+      init: jest.fn(),
     },
   },
   CLINoFlowReport: {
     instance: jest.fn(() => ({
       setIsHeadless: jest.fn(),
-    }))
+    })),
   },
 }));
 jest.mock('../app-config');
@@ -42,18 +42,20 @@ describe('test attachUsageData', () => {
   mockContext.pluginPlatform = new PluginPlatform();
   mockContext.pluginPlatform.plugins.core = [new PluginInfo('', version, '', new PluginManifest('', ''))];
   mockContext.usageData = {
-    init : jest.fn(),
-    setIsHeadless : jest.fn(),
-    emitError : jest.fn(), 
-    emitAbort : jest.fn(), 
-    emitSuccess : jest.fn(),
-    startCodePathTimer : jest.fn(), 
-    stopCodePathTimer : jest.fn(), 
-    pushHeadlessFlow : jest.fn(), 
-    pushInteractiveFlow : jest.fn(), 
-    getFlowReport : jest.fn(), 
-    assignProjectIdentifier : jest.fn(),
-  }
+    init: jest.fn(),
+    setIsHeadless: jest.fn(),
+    emitError: jest.fn(),
+    emitAbort: jest.fn(),
+    emitSuccess: jest.fn(),
+    startCodePathTimer: jest.fn(),
+    stopCodePathTimer: jest.fn(),
+    pushHeadlessFlow: jest.fn(),
+    pushInteractiveFlow: jest.fn(),
+    getFlowReport: jest.fn(),
+    assignProjectIdentifier: jest.fn(),
+    getUsageDataPayload: jest.fn(),
+    calculatePushNormalizationFactor: jest.fn(),
+  };
 
   const stateManagerMocked = stateManager as jest.Mocked<typeof stateManager>;
   stateManagerMocked.metaFileExists.mockReturnValue(true);
@@ -116,6 +118,5 @@ describe('test attachUsageData', () => {
       {},
       ts,
     );
-
   });
 });