+++ packages/amplify-cli/src/domain/amplify-usageData/IUsageData.ts
@@ -1,5 +1,6 @@
 import { IFlowData } from 'amplify-cli-shared-interfaces';
 import { Input } from '../input';
+import { UsageDataPayload } from './UsageDataPayload';
 
 /**
  * Base interface for emitting usage data
@@ -16,12 +17,14 @@ interface IUsageMetricsData {
     projectSettings: ProjectSettings,
     processStartTimeStamp: number
   ) => void;
+  getUsageDataPayload: (error: Error | null, state: string) => UsageDataPayload
   startCodePathTimer: (codePath: StartableTimedCodePath) => void;
   stopCodePathTimer: (codePath: StoppableTimedCodePath) => void;
+  calculatePushNormalizationFactor: (events: { StackId: string, PhysicalResourceId: string } [], StackId: string) => void;
 }
 
 /**
- *
+ * Interface for UsageData
  */
 export interface IUsageData extends IUsageMetricsData, IFlowData {
 }