+++ packages/amplify-cli/src/domain/amplify-usageData/NoUsageData.ts
@@ -4,11 +4,42 @@ import { IFlowReport } from 'amplify-cli-shared-interfaces/lib/amplify-cli-flow-
 import { ICommandInput, IFlowData } from 'amplify-cli-shared-interfaces';
 import { CLINoFlowReport } from './NoFlowReport';
 import { IUsageData } from './IUsageData';
+import { UsageDataPayload } from './UsageDataPayload';
+import { Input } from '../input';
 
 /**
  * Noop implementation of IUsageData used when customers have usage data turned off
  */
 export class NoUsageData implements IUsageData, IFlowData {
+  /**
+   * Return a default payload
+   * @param error 
+   * @param state 
+   * @returns 
+   */
+  getUsageDataPayload(_er: Error | null, state: string): UsageDataPayload  {
+    return new UsageDataPayload(
+      "",
+      "",
+      "",
+      new Input([]),
+      _er,
+      state,
+      "",
+      {},
+      {},
+      {},
+      { version: "", category: "", cmd: "", executable: "", input: { argv: [] }, isHeadless: true, optionFlowData: [], runtime:"", subCmd: "", timestamp: "" },
+    );
+  }
+  
+  /**
+   *  Noop implementation of calculatePushNormalizationFactor
+   */
+  calculatePushNormalizationFactor(__events: { StackId: string; PhysicalResourceId: string; }[], __stackId: string): void {
+    /* noop */
+  }
+
   /**
    * Noop implementation of emitError
    */
@@ -46,10 +77,10 @@ export class NoUsageData implements IUsageData, IFlowData {
   stopCodePathTimer(): void { /* noop */ }
 
   /**
-   * Noop function 
+   * Noop function
    */
   // eslint-disable-next-line class-methods-use-this
-  pushInteractiveFlow = (_prompt: string, _input: unknown): void => {
+  pushInteractiveFlow = (__prompt: string, __input: unknown): void => {
     /* noop */
   }
 
@@ -57,15 +88,15 @@ export class NoUsageData implements IUsageData, IFlowData {
    * Noop function
    */
   // eslint-disable-next-line class-methods-use-this
-  pushHeadlessFlow = (_headlessFlowDataString: string, _input: ICommandInput): void => {
+  pushHeadlessFlow = (__headlessFlowDataString: string, __input: ICommandInput): void => {
     /* noop */
   }
 
   /**
    * Noop function to set isHeadless flag in flowLogger
-   * @param _headless 
+   * @param _headless
    */
-  setIsHeadless = (_headless: boolean): void => {
+  setIsHeadless = (__headless: boolean): void => {
     /* noop */
   }