+++ packages/amplify-util-mock/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.4.5...amplify-util-mock@4.4.6) (2022-06-15)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [4.4.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.4.4...amplify-util-mock@4.4.5) (2022-06-13)
+
+
+
+# 8.5.0 (2022-06-10)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [4.4.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.4.3...amplify-util-mock@4.4.4) (2022-06-03)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
 ## [4.4.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.4.2...amplify-util-mock@4.4.3) (2022-05-27)
 
 **Note:** Version bump only for package amplify-util-mock